base solution

+++ FindTheWinners/TestData.cs
@@ -2,13 +2,13 @@
 {
     public class TestData
     {
-        public TestData(int playerId, decimal winAmount)
+        public TestData(int playerId, double winAmount)
         {
             PlayerId = playerId;
             WinAmount = winAmount;
         }
 
         public int PlayerId { get; private set; }
-        public decimal WinAmount { get; private set; } 
+        public double WinAmount { get; private set; } 
     }
 }
